model/requirements.txt
accelerate
torch
torchvision
git+https://github.com/huggingface/diffusers.git
Pillow
transformers
opencv-python
openai
gradio
einops